scripts/serp_sync/requirements.txt
# SERP Sync dependencies
# Install with: pip install -r requirements.txt

# SQL Server
pyodbc>=4.0.39

# Google Auth + Search Console API
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# Google Ads API (keyword volumes)
google-ads>=24.0.0

# Python-based scheduler (optional — alternative to Task Scheduler)
schedule>=1.2.1
